+++ src/test/java/com/factual/demo/DemoCafeSearch.java
@@ -13,13 +13,13 @@ import com.factual.driver.ReadResponse;
 public class DemoCafeSearch {
 
   public static void main(String[] args) {
-Factual factual = factual();
+    Factual factual = factual();
     Query q = new Query()
     .only("name", "address")
     .search("cafe")
     .within(new Circle(34.06018, -118.41835, 5000))
-    .field("postcode").equal("90067")
-    .field("category").equal("Food & Beverage")
+    .field("postcode").isEqual("90067")
+    .field("category").isEqual("Food & Beverage")
     .limit(25);
 
     // Run the query on Factual's "places" table